requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
pydantic-settings
sqlalchemy
alembic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
bcrypt
